python/requirements.txt
flask>=2.0.1
torch>=2.2.0
torchvision>=0.15.0
Pillow>=9.5.0
safetensors>=0.3.1 